main activity updates
