+++ app/build.gradle.kts
@@ -59,7 +59,9 @@ dependencies {
     implementation("androidx.constraintlayout:constraintlayout:2.1.3")
     implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.4.1")
     implementation("androidx.recyclerview:recyclerview:1.2.1")
-    implementation("io.coil-kt:coil:1.4.0")
+    implementation("io.coil-kt:coil:2.2.0")
+    implementation("io.coil-kt:coil-compose:2.2.0")
+    androidTestImplementation("androidx.test.ext:junit:1.1.3")
     implementation("androidx.navigation:navigation-fragment-ktx:2.4.1")
     implementation("androidx.navigation:navigation-ui-ktx:2.4.1")
     implementation("androidx.paging:paging-runtime-ktx:3.1.1")
